jc-46: a sanitizer finds what reading the code did not
UndefinedBehaviorSanitizer, on the first Linux CI run this project ever made,
reported at solution.c:291:

    left shift of 222 by 24 places cannot be represented in type 'int'

game->solutiondata[N] is an unsigned char, which integer-promotes to a SIGNED
int, so << 24 on a byte >= 0x80 overflows into the sign bit. Two sites: the RNG
seed in expandsolution() and the recorded best time in readsolution(). Both now
assemble in unsigned types -- the idiom fileio.c:308 already used. Upstream's,
on the 2.3.1 import.

Seeds are random 32-bit values, so this fired on roughly half of every solution
file ever recorded -- ordinary files, not malformed ones. No replay was ever
affected, and that is not luck: restartprng() masks with 0x7FFFFFFF and the
writer keeps only four bytes, so the damaged bits were discarded at both
consumers. The fix is for the undefined step in the middle.

Measured, not assumed:

  * corpus differential, jc-45 against jc-46 -- 289 sets, 18,640 valid and
    1,107 invalid under both, 0 of 303 per-set outputs differ
  * 10 unit runs / 17,059 checks; 12 end-to-end cases / 35 checks
  * mutation: reverting either hunk alone kills the suite with SIGILL

This is the first defect here found by a TOOL rather than by a person. Every
earlier one (jc-44's three, jc-45's one) was found by reading a parser,
suspecting a line, and hand-building a test to observe it -- a method that works
and does not scale. This was in a line nobody had reason to look at, exposed
through a test that had been green for weeks: the "222" is the 0xDE of that
test's 0xDEADBEEF seed. It validates the job added in 1b753e0 on its first run.

Sanitizers are also not as Linux-only as this repo claimed. -fsanitize=address
genuinely cannot run under mingw-w64, which ships no libasan -- but
-fsanitize=undefined with -fsanitize-undefined-trap-on-error needs no runtime
library and works on Windows today. ADR 0010 corrected; CLAUDE.md documents it.

The review of this change found three things the change itself did not:

  * package.ps1 defaulted to -Exe build-jc35\tworld2.exe, a gitignored leftover
    of the desync project, so the DOCUMENTED release command staged a binary
    reporting "build jc-35". Only the tag check and release.yml's explicit -Exe
    stood between that and a mislabeled download.
  * series_test.c compiled the wrong half of series.c for want of -DTWPLUSPLUS
    -- exercising gameseriescmp_name(), which never ships, and never exercising
    removefilenamesuffixes(), which does. CLAUDE.md section 3.3's trap, one file
    over; input_test.c had carried the flag correctly all along.
  * readsolution()'s besttime read had no coverage at all, so half of this fix
    was unpinned. A file-based case now reaches it. It cannot tell fixed from
    broken by value -- same-width int either way -- so its job is to make the
    line execute with bit 31 set and give the sanitizer something to see.

test/run-sanitizers.sh also stopped printing "clean under ASan+UBSan" for runs
that used neither, and now supplies -Dstricmp=strcasecmp off Windows, matching
CMakeLists.txt.

+++ solution.c
@@ -286,9 +286,21 @@ int expandsolution(solutioninfo *solution, gamesetup const *game)
     solution->flags = game->solutiondata[6];
     solution->rndslidedir = indextodir(game->solutiondata[7] & 7);
     solution->stepping = (game->solutiondata[7] >> 3) & 7;
-    solution->rndseed = game->solutiondata[8] | (game->solutiondata[9] << 8)
-					      | (game->solutiondata[10] << 16)
-					      | (game->solutiondata[11] << 24);
+    /* MOD (Jeremy, jc-46): assemble in unsigned. Each byte promotes to a
+     * SIGNED int, so << 24 on a byte >= 0x80 overflows into the sign bit --
+     * undefined behavior, and UndefinedBehaviorSanitizer halts on it. Seeds
+     * are random 32-bit values, so this fired on roughly half of every
+     * solution file ever recorded, not on malformed ones. The signed result
+     * also sign-extended on the way in: rndseed is unsigned long, 64 bits on
+     * LP64, so a negative int widened to 0xFFFFFFFF........ there. No replay
+     * was ever affected -- restartprng() masks with 0x7FFFFFFF and the writer
+     * below keeps only four bytes -- but it is still undefined, and a compiler
+     * is free to stop being kind about it. fileio.c:308 already reads a
+     * 32-bit field this way; this is the same idiom. */
+    solution->rndseed = (unsigned long)game->solutiondata[8]
+			    | ((unsigned long)game->solutiondata[9] << 8)
+			    | ((unsigned long)game->solutiondata[10] << 16)
+			    | ((unsigned long)game->solutiondata[11] << 24);
 
     initmovelist(&solution->moves);
     act.when = -1;
@@ -499,9 +511,18 @@ static int readsolution(fileinfo *file, gamesetup *game)
     if (size == 6)
 	return TRUE;
 
-    game->besttime = game->solutiondata[12] | (game->solutiondata[13] << 8)
-					    | (game->solutiondata[14] << 16)
-					    | (game->solutiondata[15] << 24);
+    /* MOD (Jeremy, jc-46): the same signed-shift overflow as the seed in
+     * expandsolution(), and EQUALLY undefined -- not a milder version of it.
+     * The only reason it never showed is that the result lands in an int of the
+     * same width either way, so the bits agreed by coincidence of types.
+     * Assembling in unsigned and converting back is implementation-defined
+     * rather than undefined, and gcc's documented modulo wrap reproduces
+     * exactly the bits the old expression produced. A .tws claiming a time with
+     * bit 31 set is still read as the same negative int it always was. */
+    game->besttime = (int)((unsigned int)game->solutiondata[12]
+			    | ((unsigned int)game->solutiondata[13] << 8)
+			    | ((unsigned int)game->solutiondata[14] << 16)
+			    | ((unsigned int)game->solutiondata[15] << 24));
     size -= 16;
     if (!game->number && !*game->passwd) {
 	game->sgflags |= SGF_SETNAME;
